v6: Move the response sub-components into `response-editor` and out of the public barrel
`ResponseHeader`, `ResponseTreeView`, and `ResponseTableView` are rendered only
by `ResponseEditor`, and their behavior is specific to a GraphQL response body:
the tree flattens the `data` envelope, and the table guesses which
list-of-objects to render. Exporting them would freeze those heuristics into the
6.0 contract.

They now live under `components/response-editor/`, leaving `ResponseEditor` as
the only export.
