+++ packages/graphiql-react/src/components/index.ts
@@ -36,12 +36,6 @@ export type { ConnectionStatus } from './status-bar';
 export { SidePanel } from './side-panel';
 export { ActivityRail } from './activity-rail';
 export type { ActivityRailProps } from './activity-rail';
-export { ResponseHeader } from './response-header';
-export type { ResponseHeaderProps } from './response-header';
-export { ResponseTableView } from './response-table-view';
-export type { ResponseTableViewProps } from './response-table-view';
-export { ResponseTreeView } from './response-tree-view';
-export type { ResponseTreeViewProps } from './response-tree-view';
 export { VarHeadersStrip } from './var-headers-strip';
 export type { VarHeadersStripProps, VarHeadersTab } from './var-headers-strip';
 export { SettingsDialog } from './settings-dialog';
