+++ packages/graphiql-react/src/components/response-editor/response-header/index.css
