+++ packages/graphiql-react/src/components/response-editor/response-table-view/index.tsx
