+++ packages/graphiql-react/src/components/response-editor/response-table-view/response-table-view.stories.tsx
